Second part. Added responsive design, layout of import page is common for table and cards rows, fixed height of carf, created a new component for each import LP token card
